requirements.txt
# LLM蜜罐项目依赖库
# 核心AI和API相关
openai>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0

# 配置和环境变量
python-dotenv>=1.0.0
PyYAML>=6.0

# 数据库连接
mysql-connector-python>=8.0.0

# Web框架和CORS
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-Login>=0.6.0
Flask-SocketIO>=5.3.0

# SSH服务器
paramiko>=3.0.0 

# 网络数据包处理
scapy>=2.5.0
scapy-http>=1.8.2

# 网络接口
netifaces>=0.11.0

# 时区处理
pytz>=2023.3

# uuid
uuid

# 其他工具
werkzeug>=2.3.0